circularity check in the task model

+++ engine/src/main/java/pt/ist/socialsoftware/blendedworkflow/domain/Rule.java
@@ -49,7 +49,8 @@ public class Rule extends Rule_Base {
 		setConditionModel(null);
 		setInvariantConditionGoal(null);
 		setTaskWithRule(null);
-		getCondition().delete();
+		if (getCondition() != null)
+			getCondition().delete();
 
 		deleteDomainObject();
 	}
